requirements.txt
requests==2.31.0
torch==2.2.1
accelerate==0.27.0
transformers==4.38.1
datasets==2.17.1
numpy==1.26.4
httpx[http2]
gitpython==3.1.42
